Composite Role Done

+++ Backend/Accounts/Accounts.Data/Repositories/UserCompositeRoleRepository.cs
@@ -11,12 +11,12 @@ namespace Accounts.Data
         public UserCompositeRoleRepository(AccountDbContext context) : base(context)
         {
         }
-        public IList<UserCompositeRole> GetAllUserCompositeRole(string projectId, int userId)
+        public IList<UserCompositeRole> GetAllUserCompositeRole(string projectId, long userId)
         {
             return Entities.Include(x => x.User).Include(z => z.CompositeRole).Where(x => x.ProjectId == projectId && x.UserId == userId).ToList();
         }
 
-        public UserCompositeRole GetUserCompositeRole(string projectId, string compositeRoleId, int userId)
+        public UserCompositeRole GetUserCompositeRole(string projectId, string compositeRoleId, long userId)
         {
             return Entities.Include(x => x.User).Include(z => z.CompositeRole).Where(x => x.ProjectId == projectId && x.CompositeRoleId == compositeRoleId && x.UserId == userId).FirstOrDefault();
         }
